revert lp_core_basic example and add lp_core_uart and uart examples
