+++ esp32c6-lp-hal/examples/uart.rs
@@ -1,13 +1,14 @@
+//! Uses `LP_UART` and logs "Hello World from LP Core".
+//! Uses GPIO4 for RX and GPIO5 for TX. GPIOs can't be changed.
+//! It is neccessary to use Serial-Uart bridge connected to TX and RX to see
+//! logs from LP_UART. Make sure the LP RAM is cleared before loading the code.
+
 #![no_std]
 #![no_main]
 
 use core::fmt::Write;
 
-use esp32c6_lp_hal::{
-    delay::Delay,
-    prelude::*,
-    uart::{LpUart},
-};
+use esp32c6_lp_hal::{delay::Delay, prelude::*, uart::LpUart};
 use panic_halt as _;
 
 #[entry]
